# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EcoLibrary API",
        "description": "API de gestion d'une bibliothèque — catégories, livres, disponibilité et état de la collection.",
        "contact": {
            "email": "admin@ecolibrary.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://project.test/api/v1",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
        "/api/admin/stats": {
            "get": {
                "tags": [
                    "Administration"
                ],
                "summary": "Statistiques globales de la bibliothèque",
                "operationId": "9aca7d759e1013283b1ac06be9bfc586",
                "responses": {
                    "200": {
                        "description": "Statistiques",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_livres": {
                                            "type": "integer",
                                            "example": 150
                                        },
                                        "total_categories": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total_exemplaires": {
                                            "type": "integer",
                                            "example": 450
                                        },
                                        "exemplaires_dispo": {
                                            "type": "integer",
                                            "example": 380
                                        },
                                        "exemplaires_degrades": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "taux_disponibilite": {
                                            "type": "string",
                                            "example": "84.44%"
                                        },
                                        "livres_populaires": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Livre"
                                            }
                                        },
                                        "stats_par_categorie": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/livres/degrades": {
            "get": {
                "tags": [
                    "Administration"
                ],
                "summary": "Rapport des livres dégradés",
                "operationId": "f0b779f9d358d18fb83c1c669a617f2f",
                "responses": {
                    "200": {
                        "description": "Livres avec exemplaires dégradés, triés du plus dégradé au moins"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Inscrire un nouvel utilisateur",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Fatiha Benali"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "fatiha@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "lecteur"
                                        ],
                                        "example": "lecteur"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Se connecter",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "fatiha@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie, retourne le token"
                    },
                    "422": {
                        "description": "Identifiants incorrects"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Se déconnecter",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Profil de l'utilisateur connecté",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "Informations de l'utilisateur"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Livre": {
                "required": [
                    "categorie_id",
                    "titre",
                    "auteur"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "categorie_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "titre": {
                        "type": "string",
                        "example": "Le Petit Prince"
                    },
                    "auteur": {
                        "type": "string",
                        "example": "Antoine de Saint-Exupéry"
                    },
                    "isbn": {
                        "type": "string",
                        "example": "978-2-07-040850-4"
                    },
                    "description": {
                        "type": "string",
                        "example": "Un conte poétique et philosophique"
                    },
                    "total_exemplaires": {
                        "type": "integer",
                        "example": 5
                    },
                    "exemplaires_dispo": {
                        "type": "integer",
                        "example": 3
                    },
                    "exemplaires_degrades": {
                        "type": "integer",
                        "example": 1
                    },
                    "nb_consultations": {
                        "type": "integer",
                        "example": 42
                    },
                    "date_publication": {
                        "type": "string",
                        "format": "date",
                        "example": "1943-04-06"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Entrez votre token Sanctum : Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}